Refactored for greater OO, fixed up some unit tests. Further work needed to meet standards compliance.
