+++ templates/CRM/Phonenumbervalidator/Page/PhoneNumberValidator.css
@@ -1,0 +1,19 @@
+.regex_rule {
+  padding-right: 10px; 
+  padding-left: 10px; 
+  display: inline-block; 
+  vertical-align: text-top;
+}
+
+.allow_options {
+  padding-right: 0px; 
+  padding-left: 10px;
+}
+
+#invalidPhonesCountDisplay {
+  padding-top: 10px;
+}
+
+#invalidPhonesCountDisplay {
+  style='padding-top: 10px';
+}
